Rebase of xaiki/provier-opts
This include a big refactor carried out by vkj and xaiki in the Paris Hackathon
