+++ src/app/index.html
@@ -23,6 +23,7 @@
 	<script id="settings-container-tpl" src="templates/settings-container.tpl" type="text/x-template"></script>
 	<script id="header-tpl" src="templates/header.tpl" type="text/x-template"></script>
 	<script id="lang-dropdown-tpl" src="templates/lang-dropdown.tpl" type="text/x-template"></script>
+	<script id="play-control-tpl" src="templates/play-control.tpl" type="text/x-template"></script>
 	<script id="movie-detail-tpl" src="templates/movie-detail.tpl" type="text/x-template"></script>
 	<script id="show-detail-tpl" src="templates/show-detail.tpl" type="text/x-template"></script>
 	<script id="player-tpl" src="templates/player.tpl" type="text/x-template"></script>
@@ -115,6 +116,7 @@
 	<script src="lib/views/lang_dropdown.js"></script>
 	<script src="lib/views/movie_detail.js"></script>
 	<script src="lib/views/show_detail.js"></script>
+	<script src="lib/views/play_control.js"></script>
 
 	<script src="lib/views/settings_container.js"></script>
 	<script src="lib/views/init_modal.js"></script>
