+++ src/app/styl/views/app.styl
@@ -20,6 +20,8 @@ $MovieListPadding = calc(100% - 67px)
 @import 'player'
 @import 'videojs'
 
+@import 'dropdowns'
+
 @import 'main_window'
 @import 'initializing'
 @import 'title_bar'
