+++ src/app/styl/views/movie_detail.styl
@@ -24,7 +24,7 @@
         opacity: 0
         background-position-x 50%
         background-position-y -33px
-        -webkit-filter: blur(15px) brightness(0.7)
+        -webkit-filter: blur(15px) brightness(0.7)^M
         background-repeat: no-repeat
         background-size: cover
         &.fadein
@@ -57,9 +57,13 @@
         margin-top: 8vh
         width: 65vw
         padding-right: 4vw
+        display: flex
+        justify-content: space-between
+        flex-direction: column
 
         .meta-container
             height: calc(100% - 70px)
+            flex: 2
 
             .title
                 color: #FFF
@@ -206,9 +210,24 @@
                 backface-visibility: hidden
                 scrollable();
 
-        .bottom-container
+        .play-control
+            display: flex
+            flex-direction: row
+            justify-content: space-between
+            white-space: nowrap
+
             height: 70px
             line-height: 35px
+            margin: 10px
+
+            .setup-container
+              display: flex
+              flex-direction: row
+              justify-content: space-between
+
+            .toggles-container
+              display: flex
+              flex: 2
 
             .favourites-toggle
                 cursor: pointer
@@ -352,155 +371,6 @@
                             opacity: 0
                         &:after
                             opacity: 1
-
-            .dropdown
-                position: relative
-                height: 26px
-                width: auto
-                float: right
-                margin-right: 2%
-                color: #FFF
-                font-family: $MainFont
-                font-smoothing: antialiased
-                font-size: 12px
-                line-height: 17px
-                cursor pointer
-                padding: 5px
-                padding-right: 13px
-                &.open
-                    background-color: rgba($ButtonBgActive,0.6);
-                    border-radius: 5px 5px 0 0;
-
-                .dropdown-arrow
-                    width: 0
-                    height: 0
-                    border-left: 4px solid transparent
-                    border-right: 4px solid transparent
-                    top: 10px
-                    position: absolute
-                    float: left
-                    right: 5px
-                    border-top: 5px solid #c4c6c8
-                    border-bottom: none
-                    &.down
-                        border-bottom: 5px solid #c4c6c8
-                        border-top: none
-
-            .flag-container
-                position: absolute
-                margin-right: 5.2%
-                margin-top: 26px
-                z-index: 3
-                width: auto
-                max-width: 330px
-                max-height: 72px
-                right: 0
-                overflow-x: hidden
-                overflow-y: hidden
-                display: none
-                height: auto
-                background: rgba(20, 24, 26, 0.6)
-
-                .flag
-                    height: 16px;
-
-
-            .flag
-                margin: 3px
-                background-image: url("../images/flags/none.png")
-                background-size: 24px 24px
-                background-position: center
-                position: relative
-                float: right
-                cursor: pointer
-                width: 24px
-                height: 22px
-                &.NA
-                    background-image: url("../images/flags/none.png")
-                &.toggle
-                    top: -6px
-                    border-radius: 0
-                    background-color: transparent
-                &.ar
-                    background-image: url("../images/flags/arabic.png")
-                &.hy
-                    background-image: url("../images/flags/armenian.png")
-                &.eu
-                    background-image: url("../images/flags/basque.png")
-                &.bn
-                    background-image: url("../images/flags/bengali.png")
-                &.bs
-                    background-image: url("../images/flags/bosnian.png")
-                &.pt-br
-                    background-image: url("../images/flags/brazilian.png")
-                &.bg
-                    background-image: url("../images/flags/bulgarian.png")
-                &.zh
-                    background-image: url("../images/flags/chinese.png")
-                &.hr
-                    background-image: url("../images/flags/croatian.png")
-                &.cs
-                    background-image: url("../images/flags/czech.png")
-                &.da
-                    background-image: url("../images/flags/danish.png")
-                &.nl
-                    background-image: url("../images/flags/dutch.png")
-                &.en
-                    background-image: url("../images/flags/english.png")
-                &.et
-                    background-image: url("../images/flags/estonian.png")
-                &.fa
-                    background-image: url("../images/flags/farsi.png")
-                &.fi
-                    background-image: url("../images/flags/finnish.png")
-                &.fr
-                    background-image: url("../images/flags/french.png")
-                &.de
-                    background-image: url("../images/flags/german.png")
-                &.el
-                    background-image: url("../images/flags/greek.png")
-                &.he
-                    background-image: url("../images/flags/hebrew.png")
-                &.hu
-                    background-image: url("../images/flags/hungarian.png")
-                &.it
-                    background-image: url("../images/flags/italian.png")
-                &.id
-                    background-image: url("../images/flags/indonesian.png")
-                &.lv
-                    background-image: url("../images/flags/latvian.png")
-                &.lt
-                    background-image: url("../images/flags/lithuanian.png")
-                &.mt
-                    background-image: url("../images/flags/maltese.png")
-                &.no
-                    background-image: url("../images/flags/norwegian.png")
-                &.pl
-                    background-image: url("../images/flags/polish.png")
-                &.pt
-                    background-image: url("../images/flags/portuguese.png")
-                &.ro
-                    background-image: url("../images/flags/romanian.png")
-                &.ru
-                    background-image: url("../images/flags/russian.png")
-                &.sr
-                    background-image: url("../images/flags/serbian.png")
-                &.sk
-                    background-image: url("../images/flags/slovak.png")
-                &.sl
-                    background-image: url("../images/flags/slovenian.png")
-                &.es
-                    background-image: url("../images/flags/spanish.png")
-                &.sv
-                    background-image: url("../images/flags/swedish.png")
-                &.th
-                    background-image: url("../images/flags/thai.png")
-                &.tr
-                    background-image: url("../images/flags/turkish.png")
-                &.uk
-                    background-image: url("../images/flags/ukrainian.png")
-                &.vi
-                    background-image: url("../images/flags/vietnamese.png")
 
             .button
                 position: relative
